refactor: github oauth 관련 DTO 디렉토리 변경 - GithubOAuthManager와 같은 계층으로 이동
