+++ backend/src/main/java/kr/kro/colla/auth/infrastructure/dto/GithubAccessTokenRequest.java
@@ -1,4 +1,4 @@
-package kr.kro.colla.auth.service.dto;
+package kr.kro.colla.auth.infrastructure.dto;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.AllArgsConstructor;
